rename depngn dir to core
