+++ src/core/getCompatData.ts
