+++ src/core/getDependencies.ts
