+++ src/core/getPackageData.ts
