+++ src/core/getPackageManager.ts
